microservices/bot/app/src/requirements.txt
flask
requests
gunicorn
python-intercom
giphypop
